+++ source/formats/guitar_pro/guitarproimporter.h
@@ -20,45 +20,13 @@
 
 #include <formats/fileformat.h>
 
-namespace Gp
-{
-    struct Beat;
-    struct Document;
-    struct Header;
-    struct Measure;
-}
-
-class KeySignature;
-class ScoreInfo;
-class System;
-class TimeSignature;
-class Voice;
-
-/// Imports Guitar Pro files.
+/// Imports Guitar Pro 3/4/5 files.
 class GuitarProImporter : public FileFormatImporter
 {
 public:
     GuitarProImporter();
 
-    virtual void load(const boost::filesystem::path &filename,
-                      Score &score) override;
-
-private:
-    static void convertHeader(const Gp::Header &header, ScoreInfo &info);
-    static void convertPlayers(const Gp::Document &doc, Score &score);
-    static int convertBarline(const Gp::Measure &measure,
-                              const Gp::Measure *prevMeasure,
-                              const Gp::Measure *nextMeasure, System &system,
-                              int start, int end, KeySignature &lastKeySig,
-                              TimeSignature &lastTimeSig);
-    static void convertAlternateEndings(const Gp::Measure &measure,
-                                        System &system, int position);
-    static int convertBeat(const Gp::Beat &beat, System &system, Voice &voice,
-                           int position);
-    static void convertIrregularGroupings(const std::vector<Gp::Beat> &beats,
-                                          const std::vector<int> &positions,
-                                          Voice &voice);
-    static void convertScore(const Gp::Document &doc, Score &score);
+    void load(const boost::filesystem::path &filename, Score &score) override;
 };
 
 #endif